+++ baybe/exceptions.py
@@ -1,5 +1,9 @@
 """Custom exceptions and warnings."""
 
+import pandas as pd
+from attr.validators import instance_of
+from attrs import define, field
+from typing_extensions import override
 
 ##### Warnings #####
 
@@ -9,6 +13,24 @@ class UnusedObjectWarning(UserWarning):
     A method or function was called with undesired arguments which indicates an
     unintended user fault.
     """
+
+
+@define
+class SearchSpaceMatchWarning(UserWarning):
+    """
+    When trying to match data to entries in the search space, something unexpected
+    happened.
+    """
+
+    message: str = field(validator=instance_of(str))
+    data: pd.DataFrame = field(validator=instance_of(pd.DataFrame))
+
+    def __attrs_pre_init(self):
+        super().__init__(self.message)
+
+    @override
+    def __str__(self):
+        return self.message
 
 
 class MinimumCardinalityViolatedWarning(UserWarning):
@@ -27,6 +49,10 @@ class IncompatibleSearchSpaceError(IncompatibilityError):
     A recommender is used with a search space that contains incompatible parts,
     e.g. a discrete recommender is used with a hybrid or continuous search space.
     """
+
+
+class IncompatibleSurrogateError(IncompatibilityError):
+    """An incompatible surrogate was selected."""
 
 
 class IncompatibleAcquisitionFunctionError(IncompatibilityError):