+++ baybe/objectives/pareto.py
@@ -1,53 +1,47 @@
-"""Functionality for single-target objectives."""
+"""Functionality for multi-target objectives."""
 
-import gc
 import warnings
 from typing import ClassVar
 
 import pandas as pd
 from attrs import define, field
-from attrs.validators import instance_of
+from attrs.validators import deep_iterable, instance_of, min_len
 from typing_extensions import override
 
 from baybe.objectives.base import Objective
+from baybe.objectives.validation import validate_target_names
 from baybe.targets.base import Target
-from baybe.targets.enum import TargetMode
-from baybe.targets.numerical import NumericalTarget
-from baybe.utils.dataframe import pretty_print_df, transform_target_columns
-from baybe.utils.plotting import to_string
+from baybe.utils.basic import to_tuple
+from baybe.utils.dataframe import transform_target_columns
 
 
 @define(frozen=True, slots=False)
-class SingleTargetObjective(Objective):
-    """An objective focusing on a single target."""
+class ParetoObjective(Objective):
+    """An objective handling multiple targets in a Pareto sense."""
 
-    is_multi_output: ClassVar[bool] = False
+    is_multi_output: ClassVar[bool] = True
     # See base class.
 
-    _target: Target = field(validator=instance_of(Target), alias="target")
-    """The single target considered by the objective."""
-
-    @override
-    def __str__(self) -> str:
-        targets_list = [target.summary() for target in self.targets]
-        targets_df = pd.DataFrame(targets_list)
-
-        fields = [
-            to_string("Type", self.__class__.__name__, single_line=True),
-            to_string("Targets", pretty_print_df(targets_df)),
-        ]
-
-        return to_string("Objective", *fields)
+    _targets: tuple[Target, ...] = field(
+        converter=to_tuple,
+        validator=[
+            min_len(2),
+            deep_iterable(member_validator=instance_of(Target)),
+            validate_target_names,
+        ],
+        alias="targets",
+    )
+    "The targets considered by the objective."
 
     @override
     @property
     def targets(self) -> tuple[Target, ...]:
-        return (self._target,)
+        return self._targets
 
     @override
     @property
     def n_outputs(self) -> int:
-        return 1
+        return len(self._targets)
 
     @override
     def transform(
@@ -89,19 +83,6 @@ class SingleTargetObjective(Objective):
                 )
         # <<<<<<<<<< Deprecation
 
-        out = transform_target_columns(
+        return transform_target_columns(
             df, self.targets, allow_missing=allow_missing, allow_extra=allow_extra
         )
-
-        # TODO: Remove hotfix (https://github.com/emdgroup/baybe/issues/460)
-        if (
-            isinstance(t := self._target, NumericalTarget)
-            and t.mode is TargetMode.MIN
-            and t.bounds.is_bounded
-        ):
-            out = -out
-        return out
-
-
-# Collect leftover original slotted classes processed by `attrs.define`
-gc.collect()